whisper-cli : align token timestamps with VAD ts
This commit aligns the token timestamps with the VAD timestamps when VAD
is enabled.

The motivation of this is that currently the token timestamps that are
reported in the full json output are the timestamps that whisper sees
after the VAD has processed the audio. This means that whisper only sees
possibly filtered audio and the token timestamps are related to the
filtered audio, not the original audio. For the segment timestamps we
map/align them with original timestamps but this is not currenly done
for the token timestamps which is what this commit aims to address.

Resolves: https://github.com/ggml-org/whisper.cpp/issues/3174